code/requirements.txt
streamlit>=1.51.0
pandas>=2.3.3
numpy>=2.3.4
plotly>=6.3.1
PyJWT>=2.8.0
python-dotenv>=1.0.0
